+++ Dockerfile
@@ -27,7 +27,7 @@ WORKDIR /app
 COPY LICENSE.md .
 COPY twine-upload.sh .
 COPY print-hash.py .
-COPY print-pkg-name.py .
+COPY print-pkg-names.py .
 COPY oidc-exchange.py .
 COPY attestations.py .
